requirements.txt
langchain==0.0.154
PYPDF2==3.0.1
pyton-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
streamlit-extras
